+++ CHANGELOG.md
@@ -9,6 +9,13 @@
   method_data dict parameters of **create_alert** method.
 * **create_credential** requires a credential_type argument now.
 * Optional parameters are required to be passed as keyword arguments.
+* Fixed **get_reports** sending the wrong protocol command
+* Removed **format_id** argument from get_reports
+* **get_report** method **format_id** argument got renamed to
+  **report_format_id**
+* Removed **alert_id** argument from **get_reports**
+* Added new **run_alert** method for running an alert method on a specific
+  report.
 
 ## gvm.xml